requirements.txt
gspread
python-dotenv
oauth2client
